+++ src/dimension/index.ts
@@ -1,6 +1,8 @@
 export {type Alternatives} from "./alternatives/alternatives";
 export * from "./conversion";
 export * from "./dimension";
+export * from "./assert-is-measurement";
+export * from "./is-measurement";
 export * from "./measurement";
 export * from "./to-fixed";
 export * from "./to-precision";
